fix:count, rating 변수 Long → Integer로 변경

+++ src/main/java/com/shootdoori/match/dto/TeamDetailResponseDto.java
@@ -10,7 +10,7 @@ public record TeamDetailResponseDto(
     String university,
     SkillLevel skillLevel,
     TeamType teamType,
-    Long memberCount,
+    Integer memberCount,
     String createdAt
 ) {
