+++ src/main/java/com/shootdoori/match/dto/TeamReviewRequestDto.java
@@ -3,8 +3,8 @@ package com.shootdoori.match.dto;
 public record TeamReviewRequestDto(Long matchId,
                                    Long reviewerTeamId,
                                    Long reviewedTeamId,
-                                   Long rating,
+                                   Integer rating,
                                    String comment,
-                                   Long punctualityRating,
-                                   Long sportsmanshipRating) {
+                                   Integer punctualityRating,
+                                   Integer sportsmanshipRating) {
 }
