+++ src/main/java/com/shootdoori/match/entity/Team.java
@@ -43,7 +43,7 @@ public class Team {
     private TeamType teamType = TeamType.OTHER;
 
     @Column(name = "MEMBER_COUNT", nullable = false, columnDefinition = "INT DEFAULT 0")
-    private Long memberCount = 0L;
+    private Integer memberCount = 0;
 
     @Enumerated(EnumType.STRING)
     @Column(name = "SKILL_LEVEL", nullable = false, columnDefinition = "VARCHAR(20) DEFAULT '아마추어'")
@@ -106,7 +106,7 @@ public class Team {
         return teamType;
     }
 
-    public Long getMemberCount() {
+    public Integer getMemberCount() {
         return memberCount;
     }
 
@@ -154,7 +154,7 @@ public class Team {
         }
     }
 
-    private void validateMemberCount(Long count) {
+    private void validateMemberCount(Integer count) {
         if (count < 0 || count > 100) {
             throw new IllegalArgumentException("멤버 수는 0~100명입니다.");
         }
